+++ fuc/api/pymaf.py
@@ -1811,8 +1811,8 @@ class MafFrame:
         return ax
 
     def plot_mutated_matched(
-        self, af, patient_col, group_col, group_order, ax=None, figsize=None,
-        **kwargs
+        self, af, patient_col, group_col, group_order, count=10, ax=None,
+        figsize=None, **kwargs
     ):
         """
         Create a bar plot visualizing the mutation prevalence of top
@@ -1828,6 +1828,8 @@ class MafFrame:
             AnnFrame column containing sample group information.
         group_order : list
             List of sample group names.
+        count : int, defualt: 10
+            Number of top mutated genes to display.
         ax : matplotlib.axes.Axes, optional
             Pre-existing axes for the plot. Otherwise, crete a new one.
         figsize : tuple, optional
@@ -1841,7 +1843,7 @@ class MafFrame:
         matplotlib.axes.Axes
             The matplotlib axes containing the plot.
         """
-        df = self.matrix_waterfall_matched(af, patient_col, group_col, group_order)
+        df = self.matrix_waterfall_matched(af, patient_col, group_col, group_order, count=count)
         df = df.applymap(lambda x: 0 if x == 'None' else 1)
         s = df.sum(axis=1) / len(df.columns) * 100
         s.name = 'Count'